Move custom variant validation code to Widget class
Instead of validating the custom variant when it is obtained by the
preserve logic, validate it directly in Widget.setData().

This leads to clearer stacktraces because they will include the code
that set the invalid variant. It also simplifies the code.

The caching of the custom variant in RemoveWidgetAdapter had been
introduced in order to reduce calls to the validation code. As the
validation is now checked only when setting a new value, this is no
longer needed. Remove it.

Refactor validating method and rework the tests.

Change-Id: Ic5ffd2cecea4b436f398b419f281c63866953386