Fix crash caused by hiding tooltip during disappear animation
In some cases toolTips are supposed to disappear immediately (no delay
or animation), for example when hovered. This is achieved by temporarily 
disabling the disappear animation. However, if the animation is already 
running this causes a crash. So now, simply call "skip" in this case.

Fix for Bug 445641 - [Tree][Table] Application crashes when dragging an
item with animating tooltip

Change-Id: I60b766b6cd7723c73525658a81b50f8c62eff0de
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>