Preserve active/cancel keys in a field instead of hashmap
Active and cancel keys (RWT.ACTIVE_KEYS and RWT.CANCEL_KEYS) are
preserved for every control. Storing it in a field instead of the
hashmap in ControlRemoteAdapter spares a lot of HashEntries.

These properties are also used on Display.setData(). However, only
change it for controls because Display uses a WidgetRemoteAdapter.

Change-Id: I66459935ebf1043b2b20c65a2c9d55ae5e1fad5c