Fix broken EventProxy_Test caused by bug in chrome canvas implementation
As of Chrome version 42 (and at least up to chromium version 45) the
canvas context "forgets" some values (e.g. fillstyle) if the canvas
element has no dimension (width and height are zero), causing
testPaintResetsPropsToWidgetValues to fail. Fixed by giving the test
canvas dimensions.

This should not be a real-life issue, since a canvas of the size 0x0
can't be seen anyway, and any resize also causes a redraw.

Change-Id: I1601a10b2a5a411fb087a145b718690f068eca80
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>