Avoids creating copies of bounds when preserving
Provide access to the internal property bounds in ControlAdapter and
use this property for preserving and comparing.

This will reduce load by the garbage collection and speed up the
comparison if bounds haven't changed because equals returns fast
on same instances.

We can be sure that the property is not modified by application code,
because public mutator methods create a new Rectangle.

Change-Id: Iccdcf202956ee0b615be766b0e9b64be947e69e0