Fix in Control.isTabGroup(), tests for Composite.getTabList()
When adapting the original SWT code in Control.isTabGroup() to our
old "only single return" pattern, we forgot to break from the loop
when setting result to true. Use separate returns instead.

This didn't lead to a bug as it's actually dead code: in our code base
the only consumer is Composite.getTabList() that only calls this method
when parent._getTabList() is null. Keep the implementation anyway in
case we use this method for other purposes later. Performance-wise it's
not critical.

Use the opportunity to improve unit tests for Composite.getTabList().

Change-Id: I097b743c7c861ff517a9051699f475e6fe229381