Do not set the lineHeight CSS property of input elements
Modern browser scale up web pages for high-density displays which may
cause rounding errors. As a result, setting the line-height of "input"
elements the way we do may cut off the lower parts of characters like
"g". 

The line height is no longer enforced now, with the server side of the
widgets guessing it, when necessary, to be font-size * 1.4.

Fix Bug 467727 - [Text] Some characters are cut off at the bottom in
Firefox 38

Change-Id: Iddd17b38f67419e10b3219655e97e6808dffbbf1
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>