Remove patch fragment preparation from bundles
Some bundles were prepared to be modified after build time with a patch
fragment by adding a virtual patch.jar as first entry on the bundle
classpath.

Since all RAP bundles are usually signed, this concept won't work in
most cases as long as the patch fragment changes single classes in a
package. Furthermore this is considered bad practice in OSGi.

In addition to the removal of the patch.jar from the Bundle-ClassPath,
this change removes the Eclipse-ExtensibleAPI marker that was required
by the tooling to detect this kind of classpath modification.

Change-Id: I3c48925cde1df271bc6e30de312f7d85d9a980c2
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>