Prevent crash when disposing widget with scheduled event notification
Check that a widget is not disposed before attempting to send a
selection event notification in EventUtil.js. Most widgets call
notifySelection/notifyDefeaultSelected async, so there is a chance the
widget was disposed in between, causing a crash.

Fix Bug 467509 - Date time in dialog processes javascript error when
clicking 3 times really fast

Change-Id: I3e70b8f6637da74cbdc1ecfaa8bb903437fee034
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>