Rework concurrency handling in RWTServlet
There's a clear distinction between the first UI request that doesn't
have a CID parameter and the subsequent UI requests that do. This
distinction allows us to create the UI session in a deterministic way
instead of creating it lazily.

If a UI request does not have a CID, we don't need to look for an
existing UISession, instead we create a new one.

If a UI request has a CID, we do not have to create one, we get the
UISession from the HttpSession.

Remove the synchronization on the servlet class that was only required
because the UISession was lazily created.

Change-Id: I6e9eb8515585dbea11396191af61e8f91c533396