Fix GridColumnLabel.js content width calculation
When implementing the grouped column feature for Nebula Grid, the method
"cellIsDisplayable" was overwritten in GridColumnLabel as part of a
mechanism to right-align the chevron image. This mechanism become
obsolete when the "expandFlexCell" method was introduced in
MultiCellWidget. It was removed, but "cellIsDisplayable" was
(incorrectly) still overwritten.

Until recently this did not cause any problems due to a random
implementation detail in MultiCellWidget._getContentWidth (it did not
use the method in case the hint was "skipFlexible"). With the recent
refactoring of MultiCellWidget (when vertical support was added), this
was no longer the case. The "cellIsDisplayable" incorrectly indicated
the last cell to be visible, so the spacing between the two was
subtracted from the label cell width.

Simply fix the issue by removing the overwriting method. No unit test
since I see no point in adding a test for the absence of certain code.

Fix for Bug 468481 - Text Truncated after column.pack() with
text-overflow set to cli

Change-Id: I9dae2a4ed9aa21ec745ae10da804aa2223317e81
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>