Fix session timeout handling in LifeCycleServiceHandler
When a client sends a request with an unknown connection id, the UI
session in LifeCycleServiceHandler#service will be null. Assuming a
well behaving client, this is only possible if the UI session has
timed out, i.e. it has been removed from the HTTP session.

Send a session timeout response in case of a null UI session in
LifeCycleServiceHandler#service. Remove the old code for handling
session time out.

If a custom client sends an unknown cid parameter for whatever reason,
a session timeout response is also acceptable.

Change-Id: I1b9db5f2518a191cdc3b6c8f8c30a58b593ff376