Cleanup obsolete session restart code
When the page is reloaded in the browser (F5 key is pressed), a new
request with a zero request counter is sent. Earlier versions of RAP
handled this as a special case "session restart" by terminating the
existing UISession and creating a new one in the same request.

With the introduction of CIDs for multi-tab support, this is not the
case anymore. Now the new request won't end up in the old UISession
because it doesn't have the old cid. Thus, the new request will be
equipped with an entirely new UISession anyway.

The old session is terminated by the shutdown request that is sent
before the page is reloaded. If this request is not sent for some
reason, it will be terminated by the session timeout.

Cleanup dead code that re-initialized the UISession and related tests.

Change-Id: Iaca5572ea587dc44cb6d3b56b66b43a46584519a