Preserve on demand for bounds property
Preserve the property `bounds` on demand. This removes the need for
preserving and comparing this property when it hasn't changed.

Change Fixture.preserveWidgets to also clear preserved values. This is
done to keep tests succeeding that rely no values being rendered after
calling this method. When all properties are preserved on demand, this
method will be replaced by clearPreserved().

Change-Id: I745a008f820774c296804cc5cb49cd324ad7bc19