Render destroy operations after rendering widgets
With support for reparenting, it can now happen that the old parent is
disposed in the same request than the set parent operation. When the
client receives the destroy operation, it will also destroy all
children, including those that are to be reparented. The set parent
operation will then lead to a crash because the widget has been
already destroyed.

By rendering destroy operations after all widgets, this cannot happen
anymore.

Fix bug 472298 - Crash with e4 demo
https://bugs.eclipse.org/bugs/show_bug.cgi?id=472298

Change-Id: I0d83cf369437272d51a5582fa276491b7bb169cc