Use Eclipse Platform property naming conventions in pom.xml
The Eclipse Platform uses property names like foo-bar-repo.url for a p2
repository URL, or foo-bar.version for specifying a version of a plugin.
This change helps to recognize those properties better for those
familiar with the Platform build.

In addition to that, it inlines Maven plug-in versions if they appear
only once, and are updated only infrequently.

Change-Id: I091e44acae0261ae72f4351f09f13b2e25bc72df
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>