Cancel DOM key events that are captured by another widget
The capture mechanism in EventHandler.js can not capture the actual DOM
events, only qooxdoo events. As a result a focused Text field will still
input characters, even if the key events are captured by an active menu
bar. Fix this by by canceling the DOM events in case a capture widget is
set. (This won't work if the capture widget is itself a Text widget, but
we never have that.)

See Bug 465098 - Menu and field get focus at the same time

Change-Id: I52e55f1cbe88810cd8669763f1936c3ed6a549e5
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>