Replace runtime-build.sh script by Maven functionality
De-activation of the test execution (skipTest==true) is done implicitly
by activating the 'eclipse-sign' profile.

The name of the final ZIP archive that includes the p2 repository is
derived in the same way as in the former shell script; it includes the
version number and the build type. The timestamp is now the time of the
build instead of the qualifier of the SDK feature.

Change-Id: Ic84ecd20cd0beca1fbb9fe6e2b876379ffd71692
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>