Make Menu bar be skipped in tabbing instead of swallowing the focus
The menu bar had getFocusRoot overwritten to prevent it from gaining
focus when clicked. Setting tabindex to null would not have sufficed as
this would just mean that the parent would get the focus, while the
desired behavior is that the focus stays where it is.

However, when using the tab key to change the focus this resulted in the
shells focusedChild property being set to null, from which tabbing can
not continue.

Do not overwrite getFocusRoot. Instead, set tabIndex to null (default
value) and explicitly exclude the MenuBar from mouse focus handling in
EventHandlerUtil.js.

Change-Id: Ib442810d73d16a145128c26a62842294c3dab355
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>