Revise handling of Menu Show and Hide events
SWT.Arm events are handled on the server side, they are not supported
by the clients. In order to notify Arm listeners on MenuItems, the
parent Menu must listen to SWT.Show events.

Instead of detecting the presence of Arm listeners on all menu items
during preserve and render, listen to Show events always if the Menu
is not a menu bars (SWT.BAR).

This may lead to an additional request when menus are opened. However,
since menus are not opened that frequently, this seems acceptable as
it produces less overhead than preserving and comparing Arm listeners
on all items in every request.

Change-Id: I5fe361a74adcb59f2bb968900aa4df5585525253