Revise handling of key listeners
Because of the network latency, there is no use in distinguishing
KeyUp and KeyDown events in the RAP protocol. Both events are mapped
to KeyDown.

Handle this case in a more standard way, i.e. preserving both events
without any exceptions, and only merge them in the render method.
Move helper methods to WidgetLCAUtil, next to the similar helpers
for Modify and Verify. I think this is the right place as untyped
listeners can be added to any widget.

Change-Id: Ieb6042f4ddfd1f9abb1b0ba9a1bcae250e57c6c1