Don't return scrollbar for Combo
In contrast to SWT, calling getHorizontalScrollBar on a Combo widget
actually returns a ScrollBar instance.

Remove the line that adds the H_SCROLL style to the Combo styles.
This code is not present in the SWT Constructor anymore (at least not
in GTK). Moreover, it is counter to the statement above that explains
why the SCROLL styles are cleared.

Add tests for both Combo and CCombo.

Change-Id: Ie74790654f2db5fa079ab465cd1a94a36b75a703