Handle ScrollBars as regular widgets in LCAs
ScrollBars are regular widgets, they are part of the RAP protocol.
However, they are currently treated in a special way: they are not
visited by the WidgetTreeVisitor and their properties are preserved
and rendered by ScrollBarLCAUtil methods that are in turn called by
the LCAs of their parent widgets.

As a result of this exception, the clearPreserved method was never
called on the RemoteAdapter of ScrollBars.

Create ScrollBarLCA, ScrollBarOperationHandler and remove the
ScrollBarUtil. This replaces this exceptional structure with the
default one, and it removes the need for redundant test methods for
scrollbar properties and events.

Include ScrollBars in the WidgetTreeVisitor. The purpose of this
helper is to traverse a complete widget tree, thus ScrollBars should
not be skipped.

Remove tests for scrollbars from the LCA tests for Scrollable widgets.
Those are now preserved and rendered automatically, so the tests
for ScrollBarLCA and ScrollBarOperationHandler should be sufficient.
Also remove those test methods that check for scrollbars visibility
after certain widget operations. Those things are already tested in
the widget tests.

Change-Id: I91505830477389f8dc15ddf2de41f888e60e7199