Preserve common listeners on demand
Preserve the current list of attached listeners when a listener is
added or removed for the first time in a request.

Only preserve common SWT listeners (those defined as constants in
the SWT class). As all these events have numbers below 64, they can
be preserved in a single long field.

Other listeners (such as for DND events, CTabFolder events, and
custom events) are not used frequently. It's acceptable to preserve
them in the corresponding LCAs.

Change-Id: I28a9556db68719d02fb7221e5f52d6acf1d8071e