Preserve active/cancelKeys on demand
With this change, later modifications to the array provided in setData
can not be detected anymore. We would need to keep a safe copy of the
given array to do so. However, this would break the contract of setData
to return the same instance in getData.

We may decide to keep active and cancel keys in separate fields later.
Since changing the given array is not a real use case, don't do that
now.

Change-Id: Icad3c02061a277ec41e1b337050df8f5480600b9