Disable warnings about missing default case
I often found myself adding a trailing empty "default: break;"
to switch-case statements, just to get rid of this warning. When
there is nothing to do in the default case, it seems that the empty
default statement is just unnecessary clutter, it's like enforcing
an else branch after every "else if".

For example, this statement:

    switch( key ) {
      case ENTER:
        handleEnter();
        break;
      case ESCAPE:
        handleEscape();
        break;
    }

would not benefit from an empty default branch:

      ...
      default:
        break;
    }

Change-Id: Ibb5f5891bfc4137cb72b6ce024873c3c48504e36