Restore the flat 2d look of MenuBar
MenuBar and ToolBar share theming values, and with the recent revisions
of the ToolBar look & feel the MenuBar accidentally also got the new 3D
look.

Add the "FLAT" state to MenuBar JavaScript objects to keep the 2D look.
This is a fix only for the maintenance branch. In master the themes for
ToolBar and MenuBar should be separated completely. See Bug 414950 -
[Theming][Menu] Client-side menu bar appearance is bound to "toolbar".


Change-Id: If5f66151a02857eef331186bd88805c0b5b15736