Fix JS error when Composite is created with H/V_SCROLL style flags
With change 53444 the ScrollBar is rendered by own LCA when parent has
V/H_SCROLL style set. On the client the ScrollBar instance is obtained
from the scrollable widget in ScrollBarHandler.js#factory and not
created directly. This is valid for all "real" scrollable widgets as
Grid, List etc. If Composite is created with these style flags the
client crashes as getHorizonta/VerticallBar functions are missing.

Create new ScrollBar instance in ScrollBarHandler.js#factory if parent
has no getHorizonta/VerticallBar functions.

This approach is not only a workaround for the bug, but allows as to
support in the future custom scrollable widgets. 

475599: JavaScript error when using RAP GEF port with RAP 3.1
https://bugs.eclipse.org/bugs/show_bug.cgi?id=475599

Change-Id: Ib8cebd4aaa56d1b05399af2eff1838e3f18ee8ab