Use singleton instances for LCAs in nebula.grid
Since LCAs are stateless by definition, we can simply return singleton
instances in getAdapter. This decouples the nebula.grid bundle from the
(now internal) LCA lookup algorithm in RWT which is going to change.

For every widget in nebula.grid, add a test to ensure that getAdapter
returns the correct LCA.

Change-Id: Ia1e3c5e0bd5b645abdf1fea4f51a8076cf219f6d