Mirror control position if parent is oriented RTL
In SWT when parent orientation is set to RIGHT_TO_LEFT the OS is
instructed to mirror the coordinate system (left become right). Thus,
without changing the layout managers, the layout is mirrored.

Introduce a "mirror" property in CanvasLayoutImpl. Based on this
property we render the left value as right property.
This is done for Composite, Shell and Group.
Change children orientation property when parent orientation is changed
by the setter.
Reverse the order of Shell caption bar sub widgets in case of RTL
orientation.

Change-Id: I77b7036b8870a2751b6ff330a9ce25d489eb8609