Fix FocusIndicator position when RTL orientation is used
When RTL orientation is set on a Button (MultiCellWidget), the right CSS
property is used on the cells instead of left. As left property is not
set in this case FocusIndicator position is calculated wrong.

Uese focus element offsetLeft property when focus indicator bounds are
calculated instead of style.left.

Change-Id: I0faad0ad51347ef83003ac5ec29f63fc33453ed4