Move preserve/renderBounds methods to remote adapter
In an effort to get rid of LCA utils, move the preserve and render
methods for the bounds property to the RemoteAdapter.

This saves some code and indirection, such as getting the remote
adapter and control adapter for a widget again and again. The remote
adapter doesn't need to expose hasPreserved and getPreserved methods
for all properties anymore.

Move tests for the generic markPreserved() and hasPreserved() methods
to the test for WidgetRemoteAdapter, where they are implemented.

Keep the tests for renderBounds in ControlLCAUtil_Test for now, as
they provide additional safety. Once ControlLCAUtil is removed,
they can be transferred into some kind of integration test.

Change-Id: I337a912158c1a50cab0f01e837e6d1783b64ea08