Fix missing scrollbars in List
With change 53444 ScrollBars are created and rendered with their own
LCA. The implementation in WidgetTreeVisitor visits the ScrollBars only
for Composite (and subclasses), but not for Scrollbable. As List
directly extends Scrollable its scrollbars are not rendered.

Change the implementation in WidgetTreeVisitor to visit scrollbars of
all scrollable widgets, not only of Composite.

https://git.eclipse.org/r/#/c/53444/

478692: [List] Does not have srollbars anymore
https://bugs.eclipse.org/bugs/show_bug.cgi?id=478692

Change-Id: I019317347c5aeb6a9f10b11c5eaf60eb5e0d6691