Enable ProgressMonitorDialog in JEE mode
ProgressMonitorDialog uses ModelContext which cannot be supported in
JEE mode as it blocks the UI thread by implementing an event loop.

For compatibility with existing code, support this dialog in JEE mode
by replacing ModalContext with a simplified background runner for
IRunnableWithProgress.

This changes the contract of the ProgressMonitorDialog's `run` method
in that this method does not block, but returns immediately.
As a consequence, exceptions that occur in the given runnable cannot
be thrown by the `run` method anymore. The application has to ensure
that all exceptions are handled within the given runnable.

Fix bug 380650: ProgressMonitorDialog does not support JEE mode
https://bugs.eclipse.org/bugs/show_bug.cgi?id=380650

Change-Id: I88ae5787f8f75e8320aceb1e76e067cf5b942cbe