Use AtomicReference in lifecycle tests
The "one-element-array" hack has been used prior to Java 5 to allow
for final variables that still can change. It's a confusing to read
as the purpose of the array is not always clear.

The AtomicXxx types are have the additional benefit of thread-safety
that might help to avoid race conditions in tests.

Change-Id: Ieb6deca5b07ab135d064e0cab80789c2e006a18b