Move preserve/renderChildren methods to remote adapter
In an effort to get rid of LCA utils, move the preserve and render
methods for control properties to the RemoteAdapter.

Keep the tests in ControlLCAUtil_Test for now. Once ControlLCAUtil is
removed, they can be transferred into some kind of integration test.

Move method to translate widget array to JSON to JsonMapping.

Change-Id: Ibb7e54f728f6bf35ca5c355b9e09f1737df2556f