Introduce UIThreadListener
Introduce UIThreadListener as a replacement for PhaseListener.
The listener methods are notified in the UI thread before and after
PROCESS_ACTION. This ensures that any changes to widgets will still
be rendered.
All other phases are not relevant for application code anymore, they
may even be removed at some point.

Change-Id: Ie4160e2b78ccbe7d712135f62c072c6230b2d800