Simplify WidgetTreeVisitor
WidgetTreeVisitor had two `visit` methods, one for Composites and one
for all other widgets. This distinction was not needed, almost all
users of this API extended AllWidgetTreeVisitor which combined
both methods in a single one.

Instead of two methods, provide only a single one, and remove
AllWidgetTreeVisitor. Turn WidgetTreeVisitor into an interface.

Move the static `accept` method to a util function since as it was
confusing to call it on the visitor. In the visitor pattern, accept
is a method of the visited node (the widget in our case) not of
the visitor.

Change-Id: Icac820c39bb9727b8163aac2d0662b54a8a2492d