Clear widget packed state after text size recalculation
When pack() method is called, the internal "packed" flag is set. It's
needed re-pack the packed widget after text size measurement. As this
flag is valid only for the current pack call, it must be reset after
text size recalculation. Otherwise the widget will be re-packed
(unnecessary) again with next text size recalculation.

475683: Shells receive pack() calls when not wanted
https://bugs.eclipse.org/bugs/show_bug.cgi?id=475683

Change-Id: I4c0d7b9c771b7aad2d83c7d17b04645fb812abf9