Render vertical gridlines for each Tree/Table/Grid cell individually
This is a precondition for implementing colspan. Now a cell can be
layouted within a row in any way and the gridline will always match. For
this change it was necessary for the rtl/mirror property to be available
outside of rendering an item, therefore setMirror was introduced in
GridRows.js. Since GridRow is not a widget and handles very low level
operations the name seemed more appropriate than setRTL.

Change-Id: I29f7eff751da897884ab34d9987e7540ae2ad20b
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>