Render re-parenting before widget disposal
Dispose operations must be rendered after changing the parent to avoid
client crashes when the old parent is disposed (see bug 472298). On the
other hand, disposed operations must be rendered before the creation of
new widgets, to respect the correct position of the widgets (when they
are added at a specific index).

Introduce ReparentedControls class that holds all reparented controls,
similar to DisposedWidgets. Render the parent property of reparented
controls in DisplayLCA (before widgets disposal) instead in
ControlLCAUtil.

481987: Separators are displayed at wrong position in context menu
https://bugs.eclipse.org/bugs/show_bug.cgi?id=481987

Change-Id: I4bc7315314b1283a2f44b29a605a34ce1de7f2c1