Favor named functions over variable assignments in JS
Named functions (`function foo() {...}`) are safer than functions
assigned to variables (`var foo = function() {...};`) because they
are "hoisted" to the beginning of the scope [1] and are therefore
always available. Moreover, this syntax is slightly shorter.

Disable JSHint warnings for functions that are used before they are
defined. These warnings aren't useful for the reasons mentioned above.

[1] http://www.adequatelygood.com/JavaScript-Scoping-and-Hoisting.html
[2] http://jshint.com/docs/options/#latedef

Change-Id: Icc698100b35951c9e6f88948b8dacf8b544daf36