Fix row template cells right alignment when vertical scrollbar appears
This was a two-part problem that was introduced by making the ScrollBars
work as overlays in RAP 3.0

1 - The rows in the grid extended beyond the client area of the grid. 

This is as side-effect of the revised row width calculation based on the
column model. When the vertical scrollbar is visible, the rows are made
a bit wider than the sum of all columns to be able to scroll far enough
horizontally. If we didn't do that the vertical scrollbar would cover
part of the last column. The row width has no effect on how the column
model is rendered within the row.

In case of templates the width of the rows is supposed to always match
the width of the grid, and horizontal scrolling is never enabled.
However, the code for the row width calculation still added the extra
area for horizontal scrolling as if the server would enable it. Now it
is only added when horizontal scrolling is actually enabled.

2 - The bounds given to the template renderer didn't respect the area
covered by the vertical scrollbar. Substract that area from the given
bounds and the renderer shifts the right-aligned cells a bit.

See Bug 481320 - RowTemplates with right aligned cell and vscroll
positioned off right edge 

Change-Id: Id6b302b2ef59ba4a42ecbcb01738dd38df4b9ed6