Fix rwt.html.Style.get not working with detached elements
A while ago we started using rwt.html.Style.get (via RWTQuery) in Grid,
but there was a problem: In some browser (at least Chrome) the result
would always be an empty string when the element was not (yet) attached
to the DOM. That resulted in broken layouts in some cases.

The method originally supported different modes, but was drastically
reduced since we adopted qooxdoo. That is what now allows us to merge it
with rwt.html.Style.getStyleProeprty. The new method works with detached
elements as well.

Fixes Bug 482823 - [Row Template] Row templates displayed in wrong way
after f5 in google chrome 


Change-Id: Ice9004f30dcd413501081414d56756dc851cd264
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>