Check currently hovered grid row before showing a grid cell tooltip
If between requesting cell tool tip text from the server and the
receiving the response the users somehow manages to move the mouse out
of the row but still keeping it inside GridRowContainer, the application
was crashing at getToolTipTargetBounds since _hoverRow is null.

Fix this by checking the hoverItem before showing the tooltip in
GridCellToolTipSupport.showToolTip. Added some comments to make the code
a little easier to understand next time we work on it.

See Bug 482338 - GridRowContainer.getToolTipTargetBounds can be called
when _hoverRow is null

Change-Id: I688e4d0a85c619f0843a199a73f2da16f6ba540d
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>