Remove GridRowContainerWrapper._getColumnOrder
The result of this method is the same as the newly introduced
gridConfig.cellOrder array. Using that is better than recalculating the
values in an error-prone way based on offsets. 

Rename the local columnOrder variables to cellOrder to match the naming
in other Grid files (cell -> index of column in grid item, column ->
actual column object).

Fix some tests that failed due to missing cellOrder value. One test
value had to be corrected since it's based on a inconsistent
(intermediate) state where the new implementation actually produces a
different outcome (GridTest line 720).

Change-Id: Ide0bea2742b74784ff36024613039f37109f0be2
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>