Remove default value initialization from GridItem.js constructor
There is no need to keep all the default values of GridItem.js in memory
all the time. Most of the time only one or two of these are actually
set, so instead of initializing them all when the item is created,
simply return the default value in the getter in case the value is
unset.

Change-Id: I4b7f2ccc6dd93c5ceec78ba7c775787eed0f37a6
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>