Fix wrong weekday is shown for 29th February of leap years
A new temporary JS Date object is created in
DateTimeDate.js#_setWeekday. It points to the current date initially (in
our case non-leap year of 2015). If you try to set day and month before
the year, 29th of February is invalid date for 2015 and it is ignored.
This leads to a wrong date and corresponding weekday.

When calculating the weekday (DateTimeDate.js#_setWeekday) we should set
year, month, date in this order to avoid invalid date (29th February of
non leap year).

483833: Weekday on 29th February of leapyears shown wrong
https://bugs.eclipse.org/bugs/show_bug.cgi?id=483833

Change-Id: I9c515e873af61ec39212d1cbb4bfc976897ae942