Use semi-transparent row overlay in Grid/Tree/Table theming
As a side-effect of recent refactorings the vertical grid lines in
Tree/Table/Grid are no longer placed on top of the rows but inside them
as part of the cell background element. As a result the overlay element
hides the the vertical lines when the row is hovered or selected, which
looks very strange, especially on hover. 

Using a semi-transparent row overlay not only fixes this issue but also
allows cell backgrounds to shine through a selection and gives the grid
a more spreadsheed-like appearance. 

Some entries in Tree.Test.css and Grid.Test.css were incorrectly
referencing Table. This was also fixed.

Change-Id: I083c91bce261c40f7ec346dd8ef9d3046a97d72d
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>