Fix Exceptions on installing/uninstalling bundles
Installing and starting a bundle which causes a RegistryChangeEvent 
from a RAP application displays an ugly exception in this application. 
The event needs the current Display but is processed in a backround
thead.

The fix is to keep a reference to the current Display from UI Thread and
reuse it in the listener executed in background thread to avoid NPE or
"IllegalStateException: No context available outside 
of the request service lifecycle"

Change-Id: I3afee116e558fa32da85c8724a7f07dd9ba8b010
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=399887
Signed-off-by: <a_mergey@yahoo.fr>