Use Java 8 for demos and tests
Launching snippets and demos from the Eclipse IDE using Java 7 leads
to UnsupportedClassVersionError because the built-in Jetty requires
Java 8.

Unit tests can as well run on Java 8. Since the build checks for
compliance with the class library API, there is no advantage in
testing with Java 7.

Since the trend is towards Java 8 anyway, update BREE and classpath
on demo and test bundles to Java 8 to avoid this kind of errors.

Keep the runtime itself compatible to Java 7.
Also keep the design.example bundle compatible to Java 7 as it is
likely used in applications.

Update launch configs and clientbuilder project.

Change-Id: I904422deed8f27520a46f967516505f5861f487d