Update versions and copyright headers for RAP 3.0.2 (Mars.2)
All RAP Runtime versions are incremented for this SR2 to 3.0.2 except
the RAP Equinox target feature that aggregates RAP Runtime and the
required bundles from the Eclipse Equinox and Jetty projects. This
target feature uses version 3.0.3 as an exception because of the
intermediate SR1a release that had to be released in the meantime.

Change-Id: I8a1f82b5be75a39bca45a79032af2907b70210ab
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>