Fix failing tests with Java 8
1. testConfigurationWithThreadContextClassLoader
For some reason, the method ClassLoader.getResources(String) is called
on the mocked class loader when running on Java 8. Since it's a mock,
this method returned null, leading to an NPE.
Using a real ClassLoader instead of a mock solves the problem.

2. testRenderData_set
This tests failed, because of different JsonObject properties order when
running on Java 8.
Re-write WidgetLCAUtil_Test#testRenderData_set assertion.

Change-Id: I9c779a14c11eefdc57e36a17174bcd50c50d3faf