Set actual tabIndex attribute value in DOM element instead of 1
The current implementation always sets 1 as a value of tabIndex
property, regardless the value sent by the server. In case of equal
tabIndex values, the tab order is based on div position. This could lead
to wrong tabbibg in saome cases.
Apply the actual tabIndex value sent by the server to the DOM element.

Change-Id: I84d28132b25cb04a36e181c1354d42087d93447b