Fix key event handling in Firefox
Since Firefox 25 "keypress" event is not fired when "keydown" event is
stopped. This is the same behaviour in other browsers.

Adjust EventHandlerUtil.js#mustRestoreKeypress for Firefox < 25 and
other browsers.
Simplify the EventHandlerUtil.js#getEventPseudoTypes. The logic stays
the same, but now it's more clear what are the differences between
Firefox and other browsers.
Adjust TestUtil.js and MenemonicHandlerTest.js accordingly.

487505: Arrow navigation in context menus not working in Firefox anymore
https://bugs.eclipse.org/bugs/show_bug.cgi?id=487505

Change-Id: I4cc2cfdccc3298cd8dc493244e9f5317caee7d6c