Fix missing mouse up event outside control
The idea of ControlOperationHandler#allowMouseEvent is to suppress mouse
events witch are not fired in SWT - mouse event on control border,
Table/Tree header or Shell title/menu bars. In SWT is possible to have a
mouse up event with negative to the Control x and y (described in bug
488755 description). The implementation of
ControlOperationHandler#allowMouseEvent must suppress by default only
mouse events on Control border.

488755: Missing Event type MouseUp
https://bugs.eclipse.org/bugs/show_bug.cgi?id=488755

Change-Id: I9c6bf41f81c824c6718a5f5e5910f9a1effbe412