Show a dialog-like box for client/server errors
For some errors we show a dialog-like box with an option to restart the
application, for other a white page with details about client/server
crash without an option to restart the application. It's better to
handle all the errors in the same way. In case of client/server error,
where the details about the crash are important, we add a separate
"Details" button to the error box to show them.

Update all places on the client to use the new error box for
client/server errors.
Add additional message for client error description.

Change-Id: I09477ce16e0e3caccbbc8f62dc8b938fac1965d5
Signed-off-by: Elshad Seyidmammadov <elshad@eclipsesource.com>