Provide OSGi and p2 capabilities to detect RWT presence
Provide a capability in the bundle manifest including a version that
allows OSGi to detect the presence of RWT in the runtime, and to wire
other bundles accordingly.

Add p2 metadata that allows p2 to decide in a similar way as OSGi which
installable units need to be taken into account.

Bug 490922 - Provide capability in RWT bundle and require it in other
RAP bundles

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=490922
Change-Id: I29bab69ffedf663e95e6790ba5940f69f867ae96
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>