Fix Location#get function to work with body element in IE
When Location#get is called for body element with "scroll" mode in IE,
the returned value is not correct (the scroll offset is not taken into
account). 
In IE body.scrollTop/Left always return zero.
To fix the issue, for body element only, use
rwt.html.Viewport.getScrollTop/Left instead.

Change-Id: I7207797da1117b39a120c0e51d6f608c8c1a971a
Signed-off-by: Elshad Seyidmammadov <elshad@eclipsesource.com>