Don't use public SWT API when synchronizing common Control properties
Common Control properties (visible, enabled...) are preserved on demand
(when SWT public API setters are used). It's possible to change one of
these properties with scripting and apply them in
ControlOperationHandler. In this case, the old value should not be
preserved as it's not changed on the server.

Add common Control properties setters to ControlAdapter.
Use these setters to apply the client value instead of using the public
SWT API.
Adjust some tests accordingly.

494881: [Control] Changing visibility state using client scripting fails
https://bugs.eclipse.org/bugs/show_bug.cgi?id=494881

Change-Id: I8767132ab5beaca4541af1d32feb1811773232a6