Fix duplicate requests when setting Browser URL with widget creation
When Browser URL is set in the same request where widget is created,
it's applyed twice on the underlying iframe - once in
Iframe.js#_applyElement and once in a timeout in BrowserHandler.js#url
property handler. This leads to duplicate requests to this URL.

Sync the source manually only if browser widget is already created.

497483: Browser.setUrl() produces duplicate GET request
https://bugs.eclipse.org/bugs/show_bug.cgi?id=497483

Change-Id: I3967d76e96560dc5cdbb901876bd0828dda1b103