Fix client crash on ToolTip parent disposal
The Tooltip widget is added to the destroyableChildren list of its
parent in ToolTipHandler.js, but it's never removed as "protocolParent"
is not set (see HandlerUtil.js#_widgetDestructor). When disposing the
ToolTip widget first and then its parent, a client crash occures,
because the disposed (removed from ObjectRegistry) ToolTip is still
available in the destroyableChildren list.

Set ToolTip "protocolParent" user data in ToolTipHandler.

511037: Error due to explicit creation of a ToolTip widget
https://bugs.eclipse.org/bugs/show_bug.cgi?id=511037

Change-Id: I3b42ac516b20cbcc3dbd5d1d46e45e2610ebbe53