Implement min/max-value support for DateTime widget
DateTime widget represents wide range of date or time data that is
selectable by users. However, in some cases, range of data or time must
be limited with according minimum and maximum values. The limitation has
to be highlighted in the calendar and prevent users to select a value
that is before the defined minimum or after the defined maximum date.

Implement min/max-value support for date, time and calendar instances of
DateTime widget.

431122: [DateTime] Enable min/max-Value Support for DateTime widget
https://bugs.eclipse.org/bugs/show_bug.cgi?id=431122

Change-Id: I1bdf12b53354bec1082c648f907f84cfad71a26d
Signed-off-by: Elshad Seyidmammadov <elshad@eclipsesource.com>